% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_params}
\alias{get_params}
\title{Create a parameter object based on a parameter file}
\usage{
get_params(
  param_file = "input_files/parameters_vaccination.xlsx",
  vaccine_file = NULL,
  dose_file = NULL,
  N_age = 9,
  L = 500,
  n_vac = 3,
  n_strain = 1,
  regions = NULL,
  regional_file = "input_files/regions.csv",
  mobility_file = "input_files/contact_matrix.csv",
  vac_priority = NULL,
  initial_conditions_file = NULL,
  import_file = NULL,
  import_age_dist_file = NULL,
  add_plus_minus_neutral = FALSE
)
}
\description{
Create a parameter object based on a parameter file
}
\keyword{parameters}
